// src/app/shared/models/Authentication/loggedInuserDetails.model.ts
export interface IloggedInuserDetails {
    Firstname: string;
    LastName: string;
    Level: string;
    MobileNo: number;
    Role: number;
    SwitchRole: string;
    UserEmail: string;
    UserId: number;
    UserType: string;
    EngagementType: string;
    TenantId: number;
    EntityId:number;
}