// src/app/shared/models/employeeModel/scheduled-site-visit.model.ts
export interface IrefModel {
    ReferralId: number;
    Name: string;
    PreferredProjectId: number;
    SalesAgentAssignedId: string;
    ReferralStatus: string;
    SalesAgentAssigned: string;
}
export interface IemployeeName {
    EmployeeId: number;
    Name: string;
}
export interface Iinteractioninfo {
    Details: string;
    EntityId: number;
    UserType: string;
    interactionType: string;
}